AsyncStreamFixture cleanup
